Fix + unify getPermLevel
